Permissões do CRUD de Usuários

+++ app/Http/Controllers/Admin/UserController.php
@@ -72,7 +72,11 @@ class UserController extends Controller
      */
     public function create()
     {
-      $this->authorize('create-user');
+      if(Gate::denies('create-user')){
+        session()->flash('msg', trans('bolao.access_denied'));
+        session()->flash('status', 'error'); // success error notification
+        return redirect()->route('home');
+      }
 
         $routeName = $this->route;
         $page = trans('bolao.user_list');
@@ -97,6 +101,12 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+      if(Gate::denies('create-user')){
+        session()->flash('msg', trans('bolao.access_denied'));
+        session()->flash('status', 'error'); // success error notification
+        return redirect()->route('home');
+      }
+
       $data = $request->all();
 
       Validator::make($data, [
@@ -124,6 +134,12 @@ class UserController extends Controller
      */
     public function show($id, Request $request)
     {
+      if(Gate::denies('show-user')){
+        session()->flash('msg', trans('bolao.access_denied'));
+        session()->flash('status', 'error'); // success error notification
+        return redirect()->route('home');
+      }
+
       $routeName = $this->route;
       $register = $this->model->find($id);
 
@@ -157,6 +173,12 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+      if(Gate::denies('edit-user')){
+        session()->flash('msg', trans('bolao.access_denied'));
+        session()->flash('status', 'error'); // success error notification
+        return redirect()->route('home');
+      }
+
       $routeName = $this->route;
       $register = $this->model->find($id);
 
@@ -187,6 +209,12 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+      if(Gate::denies('edit-user')){
+        session()->flash('msg', trans('bolao.access_denied'));
+        session()->flash('status', 'error'); // success error notification
+        return redirect()->route('home');
+      }
+
         $this->authorize('edit-user');
         $data = $request->all();
 
@@ -220,6 +248,12 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+      if(Gate::denies('delete-user')){
+        session()->flash('msg', trans('bolao.access_denied'));
+        session()->flash('status', 'error'); // success error notification
+        return redirect()->route('home');
+      }
+
       if ($this->model->delete($id)) {
         session()->flash('msg', trans('bolao.registration_deleted_successfully'));
         session()->flash('status', 'success'); // success error notification
